Import Library Manager FAQ
An FAQ hosted in the `arduino/Arduino` repository's wiki:
https://github.com/arduino/Arduino/wiki/Library-Manager-FAQ
has long served as the canonical guide for the traditional process for submitting libraries to the Library Manager index.

Now that the submission process and support for the process has been moved from `arduino/Arduino` to this dedicated
repository, it makes sense to also host the FAQ content here.

This commit imports the exact unmodified content of the FAQ from
https://github.com/arduino/Arduino/wiki/Library-Manager-FAQ/819425389e31540acee60eac9d1ade894659bf57