Don't request review for PRs from maintainer
Review requests to the PR author are not allowed. Previously, there would be a spurious workflow failure for every PR
from a user in the `env.MAINTAINERS` array.

When a maintainer is submitting a PR, they can request reviews as needed, so the automated review request can simply be
skipped in this case.