Remove instructions for library re-release
We are no longer supporting the alteration of library releases. A release should be an immutable snapshot of the library.
Having multiple revisions of a library with the same identifier can cause confusion for the users and no need for a
library maintainer to do it since they can more easily make a new release, removing the old one from the repository if
absolutely necessary.

We will always provide support for removing problematic releases from the Library Manager index, and the FAQ provides
instructions for that procedure.