Remove non-existent step from submission instructions
The library submission instructions contained a step for selecting a radio button before committing. This radio button
only exists when editing a file in a repository you have write access to, which is why I encountered it while developing
the instructions.

Library submitters will not have these radio buttons and so instructions for using a non-existent UI element would
cause confusion.