Add badge to readme for "Check General Formatting" workflow
This will make the status of the CI workflow visible at a glance to the repository maintainers, making it more likely to
bring their attention to any problems that might have developed.

I forgot to add this badge at the time I added the workflow, so playing catch up now.