Add prefix to error messages in bot comments
In the event of a problem with a submission, the comments on the pull request thread. Due to the use of a matrix job to
support submissions of any number of libraries in a single PR, this might consist of multiple comments. Adding a standard
prominent prefix (:x: **ERROR:**) to all error messages will ensure that the most important part of this information is
not missed.