Add official Microsoft Azure SDK for C Arduino library. (#768)
From the library repo https://github.com/Azure/azure-sdk-for-c-arduino.

This new library maps to the new Azure SDK for C that is designed for
embedded development, published on
https://github.com/azure/azure-sdk-for-c.

This new azure-sdk-for-c-arduino library shall be prefered over the previous azure-iot-arduino
library and dependencies (azure-iot-arduino-utility,
azure-iot-arduino-protocol-http, azure-iot-arduino-protocol-mqtt,
azure-iot-arduino-socket-esp32-wifi).

For questions, please refer to maintainer of library.