Add "pending backend" repository label
Although the submission process is completely automated, other requests (e.g., library repository URL update) are still
handled manually due to requiring additional operations on the backend

The changes to the data files stored in this repository must be coordinated with the associated changes to the Library
Manager database and/or storage systems.

The dedicated "status: pending backend" repository label will be added to the issues and PRs which can be resolved only
after those external operations have been completed; in this manner indicating their status in a clear and machine
readable manner.