Replace unsupported DS1624 library URL
Hello, eveyone!

Recently a tried to use the DS1624 library. Sadly it did not work and was lacking some features (like EEPROM read/write). I forked the original repo since it does not appear to be supported anymore. I made some changes to the library to put it in an working state.
I'm proposing a change in the library URL to my fork which was tested recently with real hardware. I'm not the original author so not sure how to proceed.

Thanks!